tests/requirements-test.txt
pytest==9.0.2
pytest-cov==7.0.0
pytest-timeout==2.4.0
torf==4.3.1
docker==7.1.0
bencode.py==4.0.0
